requirements.txt
PyYAML>=6.0 # For loading YAML configuration files
requests>=2.28 # For making HTTP requests to the A1111 API
streamlit>=1.20 # For building the web application interface
